requirements.txt
aift==1.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.4
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pythainlp==5.0.4
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
